\name{cwhtool-package}
\alias{cwhtool-package}
\docType{package}
\title{
cwh's mathematical functions
}
\description{
Miscellaneous tool functions for general use
}
\details{
\tabular{ll}{
Package: \tab cwhtool\cr
Contains: \tab cwhmath cwhstring cwhstat cwhplot cwhprint cwhtool\cr
Version: \tab 4.2.0\cr
Date: \tab 2006-01-26\cr
Depends: \tab R (>= 2.0.0)\cr
BundleDescription: \tab Miscellaneous functions for general use\cr
License: \tab GPL version 2 or newer\cr
URL: \tab http://www.wsl.ch/staff/christian.hoffmann/\cr
}

Index:
\preformatted{
cbind.colnames          Add columns to a data frame, using variable
                        names.
clean.na                Clean a matrix or data frame of rows or columns
                        of containing NA.
delayt                  Waiting loop for program execution
eql                     Check on equality, including NA==NA and
                        NaN==NaN.
libs                    List all installed packages, or all functions
                        in a package
like                    Prepare new data for prediction
ls.functions            List available functions
napply                  Apply a function to the corresponding elements
                        of two lists (?)
numberof                Count the number elements that satisfy a
                        condition.
progress.meter          Monitor the progress of a repetitive
                        calculation.
remove.dup.rows         Remove duplicate rows
select.range            Select values from a vector depending on a
                        range in a second vector.
setPowerPointStyle      Set PowerPoint style
tri                     Ternary or Triangular Plots.
trigrid                 Internal cwhtool functions
waitReturn              Wait for <Return>
}
}
\author{
Christian W. Hoffmann <christian.hoffmann@wsl.ch>

Maintainer: Christian W. Hoffmann <christian.hoffmann@wsl.ch>
}
\keyword{ package }
